Ela fly by experiment - scripts used
This commit contains the scripts used on the Pi and from Matlab.  This
was something of a hack, but it worked fairly well.  See the README file
to see how the components go together.

I have recorded calibration data with the same microphones in the
anechoic chamber, these data will be added when the calibration is done.